Refactor project structure and update dependencies
- Updated project references in `Nalix.Application.csproj` and `Nalix.Rendering.csproj` to point to the new `Nalix.NetCore.csproj`.
- Modified `HandshakeOps.cs` and `AccountService` for improved object pooling and simplified response handling.
- Restructured `Command`, `CommandOpcodes`, and `Extensions` to clarify command definitions and conversions.
- Enhanced `CredentialsPacket`, `Credentials`, and `PasswordChange` with new serialization attributes and methods for credential management.
- Created `Nalix.NetCore.csproj` with updated package references to version 3.8.0.
- Streamlined constructors in `ServerListener` and `ServerProtocol` for better connection handling.
- Updated asset paths in `GraphicsConfig` and improved connection handling in `NetworkScene`.
- Overall, these changes enhance code organization, performance, and maintainability.

+++ src/Nalix.NetCore/Security/Credentials.cs
@@ -119,7 +119,7 @@ public sealed class Credentials
 
     #region APIs
 
-    public System.Int32 EstimatedSerializedLength() => Username.Length + Password.Length;
+    public System.UInt16 EstimatedSerializedLength() => (System.UInt16)(Username.Length + Password.Length);
 
     #endregion APIs
 }
